Completed build and unit test set up, failed test though
